+++ app/src/main/java/com/example/gh_coursework/domain/entity/RouteDomain.kt
@@ -1,7 +1,7 @@
 package com.example.gh_coursework.domain.entity
 
 data class RouteDomain(
-    val routeId: Int?,
+    val routeId: Long?,
     val name: String?,
     val description: String?,
     val rating: Double?,
